# wallet-manager/docs/swagger.yaml
basePath: /
definitions:
  model.CreateWalletRequest:
    properties:
      mobile:
        type: string
    required:
    - mobile
    type: object
  model.FailureResponse:
    properties:
      error:
        type: string
      message:
        type: string
    type: object
  model.ResponseWithData:
    properties:
      data: {}
      message:
        type: string
    type: object
  model.Wallet:
    properties:
      balance:
        type: number
      created_at:
        type: string
      mobile:
        type: string
      updated_at:
        type: string
    type: object
host: http://127.0.0.1:8080
info:
  contact:
    email: support@example.com
    name: API Support
  description: This is the API documentation for the Wallet Manager service.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Wallet Manager API
  version: "1.0"
paths:
  /wallet:
    get:
      consumes:
      - application/json
      description: Retrieve a list of all wallets
      produces:
      - application/json
      responses:
        "200":
          description: Wallet list retrieved successfully
          schema:
            allOf:
            - $ref: '#/definitions/model.ResponseWithData'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/model.Wallet'
                  type: array
              type: object
        "500":
          description: Failed to retrieve wallet list
          schema:
            $ref: '#/definitions/model.FailureResponse'
      summary: Get list of all wallets
      tags:
      - Wallet
    post:
      consumes:
      - application/json
      description: Create a new wallet with the provided details
      parameters:
      - description: Wallet Creation Request
        in: body
        name: wallet
        required: true
        schema:
          $ref: '#/definitions/model.CreateWalletRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Wallet created successfully
          schema:
            allOf:
            - $ref: '#/definitions/model.ResponseWithData'
            - properties:
                data:
                  $ref: '#/definitions/model.Wallet'
              type: object
        "400":
          description: Invalid request payload
          schema:
            $ref: '#/definitions/model.FailureResponse'
        "500":
          description: Failed to create wallet
          schema:
            $ref: '#/definitions/model.FailureResponse'
      summary: Create a new wallet
      tags:
      - Wallet
swagger: "2.0"